V-main/ai-monitor/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python-headless==4.9.0.80
mediapipe==0.10.9
numpy<2.0.0
Pillow==10.2.0
scipy==1.12.0
PyPDF2==3.0.1
google-generativeai==0.8.6
ultralytics==8.2.0
